feat(loaner): add label to identify loaners

+++ app/models/vehicle.rb
@@ -5,6 +5,8 @@ require 'csv'
 class Vehicle < ApplicationRecord
   paginates_per 30
 
+  scope :visible, -> { where(hidden: false) }
+
   belongs_to :model
   belongs_to :user
 
@@ -49,25 +51,25 @@ class Vehicle < ApplicationRecord
   def remove_loaners
     return if loaner
 
-    Vehicle.where(loaner: true, vehicle_id: id).destroy_all
+    Vehicle.where(loaner: true, vehicle_id: id, user_id: user_id).destroy_all
+
+    Vehicle.where(loaner: true, user_id: user_id).find_each do |loaner_vehicle|
+      loaner_vehicle.update(
+        hidden: Vehicle.where(loaner: true, model_id: loaner_vehicle.model_id, user_id: user_id, hidden: false).where.not(id: loaner_vehicle.id).exists?
+      )
+    end
   end
 
   def create_loaner(loaner)
-    # CIG only gives you one loaner of each kind
-    Vehicle.find_or_create_by(loaner: true, model_id: loaner.id, user_id: user_id) do |new_loaner|
-      new_loaner.vehicle_id = id
-      new_loaner.public = false
-      new_loaner.purchased = true
-    end
-    # once CIG fixes this replace by:
-    # Vehicle.create(
-    #   loaner: true,
-    #   model_id: loaner.id,
-    #   user_id: user_id,
-    #   vehicle_id: id,
-    #   public: false,
-    #   purchased: true,
-    # )
+    Vehicle.create(
+      loaner: true,
+      model_id: loaner.id,
+      user_id: user_id,
+      vehicle_id: id,
+      public: false,
+      purchased: true,
+      hidden: Vehicle.exists?(loaner: true, model_id: loaner.id, user_id: user_id)
+    )
   end
 
   def broadcast_update
